Move logging macros to the `logger` module
