+++ charon/src/logger.rs
@@ -52,3 +52,71 @@ pub fn initialize_logger() {
 
     builder.init();
 }
+
+/// This macro computes the name of the function in which it is called.
+/// We adapted it from:
+/// <https://stackoverflow.com/questions/38088067/equivalent-of-func-or-function-in-rust>
+/// Note that we can't define it in macros due to technical reasons
+#[macro_export]
+macro_rules! function_name {
+    () => {{
+        fn f() {}
+        fn type_name_of<T>(_: T) -> &'static str {
+            std::any::type_name::<T>()
+        }
+        let name = type_name_of(f);
+
+        // Remove the "::f" suffix
+        let name = &name[..name.len() - 3];
+
+        // Remove the CRATE::MODULE:: prefix
+        let name = match &name.find("::") {
+            Some(pos) => &name[pos + 2..name.len()],
+            None => name,
+        };
+
+        match &name.find("::") {
+            Some(pos) => &name[pos + 2..name.len()],
+            None => name,
+        }
+    }};
+}
+
+/// A custom log trace macro. Uses the log crate.
+#[macro_export]
+macro_rules! trace {
+    ($($arg:tt)+) => {{
+        use colored::Colorize;
+        let msg = format!($($arg)+);
+        log::trace!("[{}]:\n{}", $crate::function_name!().yellow(), msg)
+    }};
+    () => {{
+        use colored::Colorize;
+        log::trace!("[{}]", $crate::function_name!().yellow())
+    }};
+}
+
+/// A custom log error macro. Uses the log crate.
+#[macro_export]
+macro_rules! error {
+    ($($arg:tt)+) => {{
+        use colored::Colorize;
+        let msg = format!($($arg)+);
+        log::error!("[{}]:\n{}", $crate::function_name!().red(), msg)
+    }};
+}
+
+/// A custom log info macro. Uses the log crate.
+#[macro_export]
+macro_rules! info {
+    ($($arg:tt)+) => {{
+        use colored::Colorize;
+        let msg = format!($($arg)+);
+        // As for info we generally output simple messages, we don't insert
+        // a breakline
+        log::info!("[{}]: {}", $crate::function_name!().yellow(), msg)
+    }};
+    () => {{
+        log::info!("[{}]", $crate::function_name!().yellow())
+    }};
+}
